+++ include/mqt-core/qdmi/driver/Driver.hpp
@@ -123,9 +123,10 @@ struct DeviceLibrary {
   /// Function pointer to @ref QDMI_device_session_create_device_job.
   decltype(QDMI_device_session_create_device_job)*
       device_session_create_device_job{};
-  /// Optional function pointer to @ref QDMI_device_session_open_device_job.
-  decltype(QDMI_device_session_open_device_job)*
-      device_session_open_device_job{};
+  /// Optional function pointer to @ref
+  /// QDMI_device_session_retrieve_device_job_by_id.
+  decltype(QDMI_device_session_retrieve_device_job_by_id)*
+      device_session_retrieve_device_job_by_id{};
   /// Function pointer to @ref QDMI_device_job_free.
   decltype(QDMI_device_job_free)* device_job_free{};
   /// Function pointer to @ref QDMI_device_job_set_parameter.
@@ -273,10 +274,10 @@ public:
   auto createJob(QDMI_Job* job) -> int;
 
   /**
-   * @brief Opens an existing job for the device.
-   * @see QDMI_device_open_job
+   * @brief Retrieves an existing job for the device.
+   * @see QDMI_session_retrieve_job_by_id
    */
-  auto openJob(const char* jobId, QDMI_Job* job) -> int;
+  auto retrieveJobById(const char* jobId, QDMI_Job* job) -> int;
 
   /**
    * @brief Frees the job associated with the device.