+++ src/fomac/FoMaC.cpp
@@ -407,11 +407,12 @@ Job Device::submitJob(const std::span<const std::byte> program,
   return jobWrapper;
 }
 
-Job Device::openJob(const std::string_view jobId) const {
+Job Device::retrieveJobById(const std::string_view jobId) const {
   const std::string id{jobId};
   QDMI_Job job = nullptr;
-  qdmi::throwIfError(QDMI_device_open_job(device_.get(), id.c_str(), &job),
-                     "Opening job");
+  qdmi::throwIfError(
+      QDMI_session_retrieve_job_by_id(device_.get(), id.c_str(), &job),
+      "Retrieving job");
   return Job{job, device_};
 }